Refactoring and debug : - Added ProgramSelect and facultySelect as resusable componenets to reduce code duplication - Debugged edit profile page - Fixed chat message layout to scroll down by default - Refactoring for optimization (useMemo to only compute once)
